fix: Override "customerId" query param and header in account interceptor
On la_throttle, most of the application is now manually passing
customerId, solution, useCase from the components.  However, the
customerId being passed by the components is not in the correct format
when using the /v2/user API.  The result when using /v2/user instead of
/v1/user was that most of the sections don't load as they're getting
`customerId = saId`, not `customerId = saId_vaId`.

With this change, the interceptor intentionally overrides the customerId
query param anad header.  This can be reverted once all of the calling
components stop passing the incorrect customerId via /v2/user.