fix(framework): Use correct user role in entitlement directive
Before this change, the user object was mapping to incorrect role
strings which the entitlement directive was using, which caused areas of
the application to be hidden when they shouldn't have been.

After this change, the user object has gets the correct role string.

On my journey to fix this, I first made changes to the entitlement
directive which got rid of the dependency on the messy user object.  I
found out after the fact that this change wasn't necessary, but I like
it so keeping it here.

Closes [CSCvs40994](https://cdetsng.cisco.com/webui/#view=CSCvs40994).