fix(lifecycle): Display Learning panel contents based on highest cxLevel
Before this commit: The learning panel contents were shown/hidden based on the
cxLevel of the current user.

After this commit: The cxLevel checks are removed from ui and the http interceptor
will no longer automatically add the saId and cxLevel headers to the elearning endpoint.
Backend will automatically filter and return data for desired highestCxLevel so no action
is needed on ui.