feat(framework): Added support for VA login
Before this change, VAs were not supposed to be able to log in to the
portal (they could, but that was a bug with the role filtering).

After this change, on application intialization, invalid smart accounts
and virtual accounts are filtered out per a modified call to /v2/user,
where an array of saId, vaId pairs is passed and /v2/user returns the
valid pairs.

Furthermore, this resolves CSCvs71778, which was a bug where users
without the proper SA or VA role could log into the portal.

Lastly, I also changed some lint rules I found onerous while developing
this.

Closes CSCvs71778